Rename event-serialization component
Resolves #533

This renames the event-serialization component to event-jackson
