+++ components/event-jackson/src/main/java/org/trellisldp/event/jackson/DefaultActivityStreamService.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.event;
+package org.trellisldp.event.jackson;
 
 import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_DATES_AS_TIMESTAMPS;
 import static java.util.Optional.empty;
