+++ components/event-jackson/src/main/resources/META-INF/beans.xml
