+++ components/event-jackson/src/main/resources/META-INF/services/org.trellisldp.api.ActivityStreamService
@@ -1,0 +1,1 @@
+org.trellisldp.event.jackson.DefaultActivityStreamService
