+++ components/event-jackson/src/test/java/org/trellisldp/event/jackson/DefaultActivityStreamServiceTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.event;
+package org.trellisldp.event.jackson;
 
 import static java.time.Instant.now;
 import static java.util.Arrays.asList;
