--- components/event-serialization/src/main/resources/META-INF/services/org.trellisldp.api.ActivityStreamService
@@ -1,1 +1,0 @@
-org.trellisldp.event.DefaultActivityStreamService
