+++ notifications/kafka/build.gradle
@@ -5,7 +5,7 @@ description = 'Trellis Kafka'
 
 ext {
     moduleName = 'org.trellisldp.kafka'
-    testModules = ['smallrye.config', 'org.trellisldp.vocabulary', 'org.trellisldp.event', 'org.apache.commons.rdf.simple']
+    testModules = ['smallrye.config', 'org.trellisldp.vocabulary', 'org.trellisldp.event.jackson', 'org.apache.commons.rdf.simple']
 }
 
 dependencies {
@@ -20,7 +20,7 @@ dependencies {
     testImplementation("io.smallrye:smallrye-config:$smallryeConfigVersion")
     testImplementation("org.apache.commons:commons-rdf-simple:$commonsRdfVersion")
     testImplementation("org.mockito:mockito-core:$mockitoVersion")
-    testImplementation project(':trellis-event-serialization')
+    testImplementation project(':trellis-event-jackson')
     testImplementation project(':trellis-vocabulary')
 }
