+++ notifications/reactive/build.gradle
@@ -5,7 +5,7 @@ description = 'Trellis Notifications (Reactive Messaging)'
 
 ext {
     moduleName = 'org.trellisldp.reactive'
-    testModules = ['smallrye.config', 'org.trellisldp.vocabulary', 'org.trellisldp.event', 'org.apache.commons.rdf.simple']
+    testModules = ['smallrye.config', 'org.trellisldp.vocabulary', 'org.trellisldp.event.jackson', 'org.apache.commons.rdf.simple']
 }
 
 dependencies {
@@ -31,7 +31,7 @@ dependencies {
         exclude group: "org.jboss.spec.javax.el", module: "jboss-el-api_3.0_spec"
     }
     testImplementation("org.mockito:mockito-core:$mockitoVersion")
-    testImplementation project(':trellis-event-serialization')
+    testImplementation project(':trellis-event-jackson')
     testImplementation project(':trellis-vocabulary')
 }
 
@@ -46,7 +46,7 @@ if (project.sourceCompatibility.isJava11Compatible()) {
             jvmArgs += [
                 '--add-opens', "$moduleName/$moduleName=weld.junit5",
                 '--add-opens', "$moduleName/$moduleName=weld.core.impl",
-                '--add-opens', "org.trellisldp.event/org.trellisldp.event=weld.core.impl",
+                '--add-opens', "org.trellisldp.event.jackson/org.trellisldp.event.jackson=weld.core.impl",
             ]
         }
     }
