+++ settings.gradle
@@ -30,7 +30,7 @@ include ':trellis-app'
 include ':trellis-audit'
 include ':trellis-constraint-rules'
 include ':trellis-dropwizard'
-include ':trellis-event-serialization'
+include ':trellis-event-jackson'
 include ':trellis-file'
 include ':trellis-io-jena'
 include ':trellis-namespaces'
@@ -65,7 +65,7 @@ project(':trellis-app-triplestore').projectDir = "$rootDir/components/app-triple
 project(':trellis-audit').projectDir = "$rootDir/components/audit" as File
 project(':trellis-constraint-rules').projectDir = "$rootDir/components/constraint-rules" as File
 project(':trellis-dropwizard').projectDir = "$rootDir/components/dropwizard" as File
-project(':trellis-event-serialization').projectDir = "$rootDir/components/event-serialization" as File
+project(':trellis-event-jackson').projectDir = "$rootDir/components/event-jackson" as File
 project(':trellis-file').projectDir = "$rootDir/components/file" as File
 project(':trellis-io-jena').projectDir = "$rootDir/components/io-jena" as File
 project(':trellis-namespaces').projectDir = "$rootDir/components/namespaces" as File
